STY: move test_instrument tests to classes
